refactor: dedupe humanBytes() into format.ts
Identical 10-line implementations lived in stats.ts, doctor.ts, and
install-plan.ts; keep one exported copy in format.ts.

Refs #538 (tidying 3/10)
