+++ src/doctor.ts
@@ -1,5 +1,6 @@
 import { existsSync, statSync } from "fs";
 import { errorMessage } from "./error-utils";
+import { humanBytes } from "./format";
 import { dirname, join, sep } from "path";
 import {
   getEngineBinPath,
@@ -81,18 +82,6 @@ export interface DoctorReport {
   stats: StatsStatus | (Partial<StatsStatus> & { error: string });
   diagnosticLogs: DoctorDiagnosticLogStatus;
   env: Record<string, string | null>;
-}
-
-function humanBytes(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`;
-  const units = ["KB", "MB", "GB", "TB"];
-  let n = bytes / 1024;
-  let i = 0;
-  while (n >= 1024 && i < units.length - 1) {
-    n /= 1024;
-    i++;
-  }
-  return `${n.toFixed(n >= 100 ? 0 : 1)} ${units[i]}`;
 }
 
 function pathSummary(path: string): PathSummary {
