+++ src/format.ts
@@ -60,3 +60,16 @@ export function formatJsonOutput(
     errors === undefined ? results : { results, errors };
   return JSON.stringify(payload, null, 2) + "\n";
 }
+
+/** Render a byte count as a human-readable size (e.g. "1.5 MB"). */
+export function humanBytes(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  const units = ["KB", "MB", "GB", "TB"];
+  let n = bytes / 1024;
+  let i = 0;
+  while (n >= 1024 && i < units.length - 1) {
+    n /= 1024;
+    i++;
+  }
+  return `${n.toFixed(n >= 100 ? 0 : 1)} ${units[i]}`;
+}
